jag kan spara användare och skapa jsondelar
Jag har dock lyckats paja spelet i sig

+++ JS/server.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const fs = require('fs/promises');
-const cors = require('cors');
 const path = require('path');
+const cors = require('cors');
 
 const app = express();
 const PORT = 3000;
@@ -11,14 +11,40 @@ app.use(express.json());
 
 const userDataPath = path.join(__dirname, '../json/users.json');
 
+app.get('/users', async (req, res) => {
+  try {
+    const userData = await fs.readFile(userDataPath, 'utf-8');
+    const users = JSON.parse(userData);
+    res.json({ users: users.users });
+  } catch (error) {
+    console.error(error);
+    res.status(500).send('Internal Server Error');
+  }
+});
+
 app.post('/register', async (req, res) => {
   const { username } = req.body;
 
   try {
+    const cleanedUsername = username.replace(/\s/g, '').toUpperCase();
+
     const userData = await fs.readFile(userDataPath, 'utf-8');
+
     const users = JSON.parse(userData);
 
-    users.users.push({ username });
+    const existingUser = users.users.find(user => user.username === cleanedUsername);
+    if (existingUser) {
+      return res.status(400).send('User already exists');
+    }
+
+    const newUser = {
+      id: users.users.length + 1,
+      username: cleanedUsername,
+      rating: 1000,
+      matchHistory: []
+    };
+
+    users.users.push(newUser);
 
     await fs.writeFile(userDataPath, JSON.stringify(users, null, 2));
